feat: ПР 14, для проверки автотестами

+++ controllers/cards.js
@@ -1,12 +1,32 @@
 const Card = require('../models/card');
-const { ERROR_CODE_400, ERROR_CODE_404, ERROR_CODE_500 } = require('../utils/constants');
+const BadRequestError = require('../errors/BadRequestError');
+const InternalServerError = require('../errors/InternalServerError');
+const NotFoundError = require('../errors/NotFoundError');
+const UnauthorizedError = require('../errors/UnauthorizedError');
 
-module.exports.getCardsAll = (req, res) => {
+module.exports.getCardsAll = (req, res, next) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
     // eslint-disable-next-line no-unused-vars
-    .catch((err) => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' }));
+    .catch((err) => {
+      throw new InternalServerError({
+        message: 'Произошла ошибка.',
+      });
+    })
+    .catch(next);
+  /*
+   .catch((err) => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' })); */
 };
+
+/*
+// controllers/cards.js
+
+module.exports.createCard = (req, res) => Card.create({
+  name: req.body.name,
+  link: req.body.link,
+  owner: req.user._id // используем req.user
+});
+*/
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
@@ -14,32 +34,63 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные при создании карточки.' });
+        throw new BadRequestError({
+          message: 'Переданы некорректные данные при создании карточки.'
+        });
+        /* return res.status(ERROR_CODE_400).send({
+          message: 'Переданы некорректные данные при создании карточки.' }); */
       }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
+      throw new InternalServerError({
+        message: 'Произошла ошибка.',
+      });
+      // return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
     });
 };
 
 module.exports.deleteCard = (req, res) => {
   const id = req.params.cardId;
-
-  Card.findByIdAndRemove(id)
-    .then((card) => {
-      if (!card) {
+  const ownerId = req.user._id;
+  const userCardId = req.params.owner;
+  if (ownerId === !userCardId) {
+    throw new UnauthorizedError({
+      message: 'Карточка создана другим пользователем. У вас нет прав на её удаление.',
+    });
+    /*
+    return res.status(ERROR_CODE_401).send({
+      message: `Карточка создана другим пользователем. У вас нет прав на её удаление.`,
+    }); */
+  } else {
+    Card.findByIdAndRemove(id)
+      .then((card) => {
+        if (!card) {
+          throw new NotFoundError({
+            message: `Карточка по указанному id ${id} не найдена.`,
+          });
+        /*
         return res.status(ERROR_CODE_404).send({
           message: `Карточка по указанному id ${id} не найдена.`,
-        });
-      }
-      return res.send({ data: card });
-    })
-    .catch((err) => {
-      if (err.name === 'ValidationError' || (err.name === 'CastError' && err.path === '_id')) {
+        }); */
+        }
+        return res.send({ data: card });
+      })
+      .catch((err) => {
+        if (err.name === 'ValidationError' || (err.name === 'CastError' && err.path === '_id')) {
+          throw new BadRequestError({
+            message: `Передан некорректный id ${id} карточки.`,
+          });
+        /*
         return res.status(ERROR_CODE_400).send({
           message: `Передан некорректный id ${id} карточки.`,
+        }); */
+        }
+
+        throw new InternalServerError({
+          message: 'Произошла ошибка.',
         });
-      }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
-    });
+
+        // return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
+      });
+  }
 };
 
 module.exports.likeCard = (req, res) => {
@@ -52,17 +103,31 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res.status(ERROR_CODE_404).send({
+        throw new NotFoundError({
           message: `Передан несуществующий id ${id} карточки.`,
         });
+        /*
+        return res.status(ERROR_CODE_404).send({
+          message: `Передан несуществующий id ${id} карточки.`,
+        }); */
       }
       return res.send({ data: card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError' || (err.name === 'CastError' && err.path === '_id')) {
-        return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для постановки лайка.' });
+        throw new BadRequestError({
+          message: 'Переданы некорректные данные для постановки лайка.'
+        });
       }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
+      /* return res.status(ERROR_CODE_400).send({
+          message: 'Переданы некорректные данные для постановки лайка.' });
+      } */
+
+      throw new InternalServerError({
+        message: 'Произошла ошибка.',
+      });
+
+      // return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
     });
 };
 
@@ -70,9 +135,13 @@ module.exports.dislikeCard = (req, res) => {
   const id = req.params.cardId;
 
   if (!req.user._id) {
-    return res.status(ERROR_CODE_404).send({
+    throw new NotFoundError({
       message: 'Переданы некорректные данные для снятия лайка.',
     });
+    /*
+    return res.status(ERROR_CODE_404).send({
+      message: 'Переданы некорректные данные для снятия лайка.',
+    }); */
   }
   return Card.findByIdAndUpdate(
     id,
@@ -82,16 +151,31 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res.status(ERROR_CODE_404).send({
+        throw new NotFoundError({
           message: `Передан несуществующий id ${id} карточки.`,
         });
+        /*
+        return res.status(ERROR_CODE_404).send({
+          message: `Передан несуществующий id ${id} карточки.`,
+        }); */
       }
       return res.send({ data: card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError' || (err.name === 'CastError' && err.path === '_id')) {
-        return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для снятия лайка.' });
+        throw new BadRequestError({
+          message: 'Переданы некорректные данные для снятия лайка.'
+        });
+
+        /* return res.status(ERROR_CODE_400).send({
+          message: 'Переданы некорректные данные для снятия лайка.' });
+        */
       }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
+
+      throw new InternalServerError({
+        message: 'Произошла ошибка.',
+      });
+
+      // return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка.' });
     });
 };
